+++ python/paddle/fluid/tests/unittests/ir/inference/test_trt_reduce_mean_op.py
@@ -23,61 +23,6 @@ from paddle.fluid.core import PassVersionChecker
 from paddle.fluid.core import AnalysisConfig
 
 
-class TRTReduceSumTest(InferencePassTest):
-    def setUp(self):
-        with fluid.program_guard(self.main_program, self.startup_program):
-            data = fluid.data(
-                name="data", shape=[-1, 3, 10, 768], dtype="float32")
-            reduce_sum = fluid.layers.reduce_sum(
-                data, dim=[2, -1], keep_dim=True)
-            out = fluid.layers.batch_norm(reduce_sum, is_test=True)
-
-        self.feeds = {
-            "data": np.random.random([3, 3, 10, 768]).astype("float32"),
-        }
-        self.enable_trt = True
-        self.trt_parameters = TRTReduceSumTest.TensorRTParam(
-            1 << 30, 32, 1, AnalysisConfig.Precision.Float32, False, False)
-        self.fetch_list = [out]
-        self.dynamic_shape_params = TRTReduceSumTest.DynamicShapeParam({
-            'data': [1, 3, 8, 8]
-        }, {'data': [3, 3, 10, 768]}, {'data': [3, 3, 10, 768]}, False)
-
-    def test_check_output(self):
-        if core.is_compiled_with_cuda():
-            use_gpu = True
-            self.check_output_with_option(use_gpu, flatten=True)
-            self.assertTrue(
-                PassVersionChecker.IsCompatible('tensorrt_subgraph_pass'))
-
-
-class TRTReduceSumAllTest(InferencePassTest):
-    def setUp(self):
-        with fluid.program_guard(self.main_program, self.startup_program):
-            data = fluid.data(
-                name="data", shape=[-1, 3, 10, 768], dtype="float32")
-            reduce_sum = fluid.layers.reduce_sum(data, keep_dim=True)
-            out = fluid.layers.batch_norm(reduce_sum, is_test=True)
-
-        self.feeds = {
-            "data": np.random.random([3, 3, 10, 768]).astype("float32"),
-        }
-        self.enable_trt = True
-        self.trt_parameters = TRTReduceSumAllTest.TensorRTParam(
-            1 << 30, 32, 1, AnalysisConfig.Precision.Float32, False, False)
-        self.fetch_list = [out]
-        self.dynamic_shape_params = TRTReduceSumAllTest.DynamicShapeParam({
-            'data': [1, 3, 8, 8]
-        }, {'data': [3, 3, 10, 768]}, {'data': [3, 3, 10, 768]}, False)
-
-    def test_check_output(self):
-        if core.is_compiled_with_cuda():
-            use_gpu = True
-            self.check_output_with_option(use_gpu, flatten=True)
-            self.assertTrue(
-                PassVersionChecker.IsCompatible('tensorrt_subgraph_pass'))
-
-
 class TRTReduceMeanTest(InferencePassTest):
     def setUp(self):
         with fluid.program_guard(self.main_program, self.startup_program):
@@ -91,10 +36,10 @@ class TRTReduceMeanTest(InferencePassTest):
             "data": np.random.random([3, 3, 10, 768]).astype("float32"),
         }
         self.enable_trt = True
-        self.trt_parameters = TRTReduceSumTest.TensorRTParam(
+        self.trt_parameters = TRTReduceMeanTest.TensorRTParam(
             1 << 30, 32, 1, AnalysisConfig.Precision.Float32, False, False)
         self.fetch_list = [out]
-        self.dynamic_shape_params = TRTReduceSumTest.DynamicShapeParam({
+        self.dynamic_shape_params = TRTReduceMeanTest.DynamicShapeParam({
             'data': [1, 3, 8, 8]
         }, {'data': [3, 3, 10, 768]}, {'data': [3, 3, 10, 768]}, False)
 
@@ -118,10 +63,10 @@ class TRTReduceMeanAllTest(InferencePassTest):
             "data": np.random.random([3, 3, 10, 768]).astype("float32"),
         }
         self.enable_trt = True
-        self.trt_parameters = TRTReduceSumAllTest.TensorRTParam(
+        self.trt_parameters = TRTReduceMeanAllTest.TensorRTParam(
             1 << 30, 32, 1, AnalysisConfig.Precision.Float32, False, False)
         self.fetch_list = [out]
-        self.dynamic_shape_params = TRTReduceSumAllTest.DynamicShapeParam({
+        self.dynamic_shape_params = TRTReduceMeanAllTest.DynamicShapeParam({
             'data': [1, 3, 8, 8]
         }, {'data': [3, 3, 10, 768]}, {'data': [3, 3, 10, 768]}, False)
